Send sector divided by 4K

+++ cheeze.h
@@ -29,7 +29,7 @@
 struct cheeze_req_user {
 	int id;
 	char *buf;
-	unsigned long pos; // sector_t
+	unsigned int pos; // sector_t but divided by 4096
 	unsigned int len;
 } __attribute__((aligned(8), packed));
